# perf: optimize .child