--- a/lib/proto.js
+++ b/lib/proto.js
@@ -43,7 +43,8 @@ const {
   asChindings,
   asJson,
   buildFormatters,
-  stringify
+  stringify,
+  noop
 } = require('./tools')
 const {
   version
@@ -86,10 +87,27 @@ function child (bindings, options) {
   if (!bindings) {
     throw Error('missing bindings for child Pino')
   }
-  options = options || {} // default options to empty object
   const serializers = this[serializersSym]
   const formatters = this[formattersSym]
   const instance = Object.create(this)
+
+  if (options == null) {
+    if (instance[formattersSym].bindings !== resetChildingsFormatter) {
+      instance[formattersSym] = buildFormatters(
+        formatters.level,
+        resetChildingsFormatter,
+        formatters.log
+      )
+    }
+
+    instance[chindingsSym] = asChindings(instance, bindings)
+
+    if (this.onChild !== noop) {
+      this.onChild(instance)
+    }
+
+    return instance
+  }
 
   if (options.hasOwnProperty('serializers') === true) {
     instance[serializersSym] = Object.create(null)
@@ -222,8 +240,6 @@ function write (_obj, msg, num) {
   }
   stream.write(streamWriteHook ? streamWriteHook(s) : s)
 }
-
-function noop () {}
 
 function flush (cb) {
   if (cb != null && typeof cb !== 'function') {