--- a/lib/proto.js
+++ b/lib/proto.js
@@ -91,6 +91,10 @@ function child (bindings, options) {
   const formatters = this[formattersSym]
   const instance = Object.create(this)
 
+  // If an `options` object was not supplied, we can improve
+  // the performance of child creation by skipping
+  // the checks for set options and simply return
+  // a baseline instance.
   if (options == null) {
     if (instance[formattersSym].bindings !== resetChildingsFormatter) {
       instance[formattersSym] = buildFormatters(