--- a/lib/tools.js
+++ b/lib/tools.js
@@ -257,10 +257,10 @@ function asChindings (instance, bindings) {
 
   for (const key in bindings) {
     value = bindings[key]
-    const valid = key !== 'level' &&
+    const valid = (key.length < 5 || (key !== 'level' &&
       key !== 'serializers' &&
       key !== 'formatters' &&
-      key !== 'customLevels' &&
+      key !== 'customLevels')) &&
       bindings.hasOwnProperty(key) &&
       value !== undefined
     if (valid === true) {