```bash
benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
pino child creation          174.95 ns/iter  87.23 ns █                    
                       (63.07 ns … 5.76 µs)   3.60 µs █                    
                    ( 53.17  b …   8.62 kb) 261.17  b █▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

pino2 child creation         257.64 ns/iter 330.37 ns █                    
                       (67.99 ns … 7.37 µs)   4.49 µs █▂                   
                    ( 22.38  b …   5.82 kb) 266.57  b ██▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

```
```js
// bench.mjs

import { bench, run } from 'mitata'

import pino from './pino.js'
import pino2 from 'pino'

let logger = pino().child({ b: 'property' })
const children = []

bench('pino child creation', function () {
  children.push(logger.child({ a: 'property' }))
})

let logger2 = pino().child({ b: 'property' })
const children2 = []

bench('pino2 child creation', function () {
  children2.push(logger2.child({ a: 'property' }))
})

await run()

children[Math.random() * children.length | 0].info({ hello: 'world' })
```